refactor: move article model to the parent folder
