+++ lib/app/features/feed/data/models/article_data.dart
