+++ lib/app/features/feed/views/components/entities_list/components/article_list_item.dart
@@ -3,7 +3,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:ion/app/extensions/num.dart';
-import 'package:ion/app/features/feed/data/models/article/article_data.dart';
+import 'package:ion/app/features/feed/data/models/article_data.dart';
 import 'package:ion/app/features/feed/views/components/article/article.dart';
 import 'package:ion/app/router/app_routes.dart';
