+++ lib/app/features/feed/views/components/entities_list/components/generic_repost_list_item.dart
@@ -3,7 +3,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:ion/app/components/screen_offset/screen_side_offset.dart';
 import 'package:ion/app/extensions/extensions.dart';
-import 'package:ion/app/features/feed/data/models/article/article_data.dart';
+import 'package:ion/app/features/feed/data/models/article_data.dart';
 import 'package:ion/app/features/feed/data/models/generic_repost.dart';
 import 'package:ion/app/features/feed/views/components/article/article.dart';
 import 'package:ion/app/features/feed/views/components/feed_item/feed_item_header/repost_author_header.dart';
